+++ PokemonEncyclopedia/PokemonEncyclopedia/FeedCache/Infrastructure/CoreData/PokemonEncyclopedia.xcdatamodeld/PokemonEncyclopedia.xcdatamodel/contents
